feat: Add removeAds property to user model
This commit introduces a new `removeAds` boolean property to the user model. This will be used to track whether a user has purchased the ad-free version of the app.

The `UserService` is updated to initialize this property to `false` for all new guest users, ensuring they see ads by default. Additionally, an unused import was removed from the friends screen.